extra components, navbar, login-form, logout-button moved to components folder
