+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from "./login-form";
+import LoginForm from "./components/login-form";
 
 export default function Home() {
   return (
